+++ apps/demo/CHANGELOG.md
@@ -1,5 +1,12 @@
 # home
 
+## 0.2.40
+
+### Patch Changes
+
+- Updated dependencies
+  - @versaprotocol/react@1.9.2
+
 ## 0.2.39
 
 ### Patch Changes